@@ -7,7 +7,7 @@ import com.jfirer.jfire.core.beanfactory.BeanFactory;
 public class SelectedBeanFactory implements BeanFactory
 {
     private final ApplicationContext applicationContext;
-    private final String beanFactoryBeanName;
+    private final String             beanFactoryBeanName;
 
     public SelectedBeanFactory(ApplicationContext applicationContext, String beanFactoryBeanName)
     {
@@ -18,7 +18,7 @@ public class SelectedBeanFactory implements BeanFactory
     @Override
     public <E> E getUnEnhanceyInstance(BeanDefinition beanDefinition)
     {
-        BeanFactory beanFactory = (BeanFactory) applicationContext.getBean(beanFactoryBeanName);
+        BeanFactory beanFactory = applicationContext.getBean(beanFactoryBeanName);
         return beanFactory.getUnEnhanceyInstance(beanDefinition);
     }
 }
