@@ -2,11 +2,9 @@ package com.jfirer.jfire.core.prepare.annotation.condition.provide;
 
 import com.jfirer.baseutil.bytecode.annotation.AnnotationMetadata;
 import com.jfirer.baseutil.bytecode.annotation.ValuePair;
-import com.jfirer.baseutil.bytecode.support.AnnotationContext;
 import com.jfirer.baseutil.bytecode.support.AnnotationContextFactory;
 import com.jfirer.jfire.core.ApplicationContext;
 import com.jfirer.jfire.core.DefaultApplicationContext;
-import com.jfirer.jfire.core.bean.BeanRegisterInfo;
 import com.jfirer.jfire.core.prepare.annotation.condition.Conditional;
 import com.jfirer.jfire.core.prepare.annotation.condition.ErrorMessage;
 import com.jfirer.jfire.core.prepare.annotation.configuration.Configuration;
@@ -14,7 +12,6 @@ import com.jfirer.jfire.core.prepare.annotation.configuration.Configuration;
 import java.lang.annotation.Annotation;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import java.util.Optional;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Conditional(ConditionOnAnnotation.OnAnnotation.class)
@@ -48,12 +45,7 @@ public @interface ConditionOnAnnotation
                     errorMessage.addErrorMessage("注解:" + each + "不存在于类路径");
                     return false;
                 }
-                boolean has  = context.getAllBeanRegisterInfos().stream()
-                                                        .filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType())
-                                                                                                            .isAnnotationPresent(Configuration.class))
-                                                        .filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType())
-                                                                                                            .isAnnotationPresent(aClass))
-                                                        .findAny().isPresent();
+                boolean has = context.getAllBeanRegisterInfos().stream().filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(Configuration.class)).filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(aClass)).findAny().isPresent();
 //                for (Class<?> configurationClass : context.getConfigurationClassSet())
 //                {
 //                    AnnotationContext annotationContext = annotationContextFactory.get(configurationClass, classLoader);
@@ -63,12 +55,13 @@ public @interface ConditionOnAnnotation
 //                        break;
 //                    }
 //                }
-                if (has == false)
+                if (!has)
                 {
                     errorMessage.addErrorMessage("注解:" + each + "没有标注在配置类上");
                     return false;
                 }
-            } return true;
+            }
+            return true;
         }
     }
 }
