@@ -1,7 +1,6 @@
 package com.jfirer.jfire.core.prepare.processor;
 
 import com.jfirer.baseutil.IniReader;
-import com.jfirer.baseutil.bytecode.support.AnnotationContext;
 import com.jfirer.baseutil.bytecode.support.AnnotationContextFactory;
 import com.jfirer.jfire.core.ApplicationContext;
 import com.jfirer.jfire.core.DefaultApplicationContext;
@@ -21,20 +20,13 @@ public class PropertyPathProcessor implements ContextPrepare
     {
         ClassLoader              classLoader              = Thread.currentThread().getContextClassLoader();
         AnnotationContextFactory annotationContextFactory = DefaultApplicationContext.ANNOTATION_CONTEXT_FACTORY;
-        context.getAllBeanRegisterInfos().stream()
-               .filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType())
-                                                                   .isAnnotationPresent(Configuration.class))
-
-                .filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(PropertyPath.class))
-                .flatMap(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).getAnnotations(PropertyPath.class).stream())
-                .flatMap(propertyPath -> Arrays.stream(propertyPath.value()))
-                .forEach(path->{
-                    IniReader.IniFile iniFile = Utils.processPath(path);
-                    for (String property : iniFile.keySet())
-                    {
-                        context.getEnv().putProperty(property, iniFile.getValue(property));
-                    }
-                });
+        context.getAllBeanRegisterInfos().stream().filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(Configuration.class)).filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(PropertyPath.class)).flatMap(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).getAnnotations(PropertyPath.class).stream()).flatMap(propertyPath -> Arrays.stream(propertyPath.value())).forEach(path -> {
+            IniReader.IniFile iniFile = Utils.processPath(path);
+            for (String property : iniFile.keySet())
+            {
+                context.getEnv().putProperty(property, iniFile.getValue(property));
+            }
+        });
 //        for (Class<?> each : context.getConfigurationClassSet())
 //        {
 //            AnnotationContext annotationContext = annotationContextFactory.get(each, classLoader);
