@@ -1,7 +1,7 @@
 package com.jfirer.jfire.test.function;
 
-import com.jfirer.jfire.core.DefaultApplicationContext;
 import com.jfirer.jfire.core.ApplicationContext;
+import com.jfirer.jfire.core.DefaultApplicationContext;
 import com.jfirer.jfire.core.inject.notated.CanBeNull;
 import org.junit.Assert;
 import org.junit.Test;
@@ -12,6 +12,43 @@ import java.util.List;
 
 public class DiTest
 {
+    /**
+     * 测试List注入
+     */
+    @Test
+    public void test()
+    {
+        ApplicationContext context = new DefaultApplicationContext();
+        context.register(ForDi1.class);
+        context.register(ForDi2.class);
+        context.register(Holder.class);
+        Holder holder = context.getBean(Holder.class);
+        Assert.assertEquals(2, holder.f.size());
+    }
+    /**
+     * 接口带有名称的注入
+     */
+    @Test
+    public void test_2()
+    {
+        ApplicationContext context = new DefaultApplicationContext();
+        context.register(ForDi1.class);
+        context.register(Holder2.class);
+        Holder2 holder = context.getBean(Holder2.class);
+        Assert.assertNotNull(holder.di);
+    }
+    /**
+     * 接口没有实现，但是允许为空
+     */
+    @Test
+    public void test_3()
+    {
+        ApplicationContext context = new DefaultApplicationContext();
+        context.register(Holder3.class);
+        Holder3 holder = context.getBean(Holder3.class);
+        Assert.assertNull(holder.di);
+    }
+
     public interface forDi
     {
 
@@ -33,7 +70,7 @@ public class DiTest
     public static class Holder
     {
         @Resource
-        private List<forDi> f = new LinkedList<DiTest.forDi>();
+        private final List<forDi> f = new LinkedList<DiTest.forDi>();
     }
 
     @Resource
@@ -49,44 +86,5 @@ public class DiTest
         @CanBeNull
         @Resource(name = "di")
         private forDi di;
-    }
-
-    /**
-     * 测试List注入
-     */
-    @Test
-    public void test()
-    {
-        ApplicationContext context = new DefaultApplicationContext();
-        context.register(ForDi1.class);
-        context.register(ForDi2.class);
-        context.register(Holder.class);
-        Holder holder = context.getBean(Holder.class);
-        Assert.assertEquals(2, holder.f.size());
-    }
-
-    /**
-     * 接口带有名称的注入
-     */
-    @Test
-    public void test_2()
-    {
-        ApplicationContext context = new DefaultApplicationContext();
-        context.register(ForDi1.class);
-        context.register(Holder2.class);
-        Holder2 holder = context.getBean(Holder2.class);
-        Assert.assertNotNull(holder.di);
-    }
-
-    /**
-     * 接口没有实现，但是允许为空
-     */
-    @Test
-    public void test_3()
-    {
-        ApplicationContext context = new DefaultApplicationContext();
-        context.register(Holder3.class);
-        Holder3 holder = context.getBean(Holder3.class);
-        Assert.assertNull(holder.di);
     }
 }
