@@ -1,7 +1,7 @@
 package com.jfirer.jfire.test.function;
 
-import com.jfirer.jfire.core.DefaultApplicationContext;
 import com.jfirer.jfire.core.ApplicationContext;
+import com.jfirer.jfire.core.DefaultApplicationContext;
 import com.jfirer.jfire.core.inject.notated.PropertyRead;
 import com.jfirer.jfire.core.prepare.annotation.PropertyPath;
 import com.jfirer.jfire.core.prepare.annotation.configuration.Configuration;
@@ -15,6 +15,28 @@ public class PropertyPathImporterTest
 {
     @PropertyRead
     private int age;
+    /**
+     * 使用classpath路径读取
+     */
+    @Test
+    public void test()
+    {
+        ApplicationContext context = new DefaultApplicationContext(Test1.class);
+        context.register(PropertyPathImporterTest.class);
+        PropertyPathImporterTest test = context.getBean(PropertyPathImporterTest.class);
+        Assert.assertEquals(12, test.age);
+    }
+    /**
+     * 使用文件路径读取
+     */
+    @Test
+    public void test2()
+    {
+        ApplicationContext context = new DefaultApplicationContext(Test2.class);
+        context.register(PropertyPathImporterTest.class);
+        PropertyPathImporterTest test = context.getBean(PropertyPathImporterTest.class);
+        Assert.assertEquals(12, test.age);
+    }
 
     @Configuration
     @PropertyPath("classpath:propertiestest.properties")
@@ -28,29 +50,5 @@ public class PropertyPathImporterTest
     public static class Test2
     {
 
-    }
-
-    /**
-     * 使用classpath路径读取
-     */
-    @Test
-    public void test()
-    {
-        ApplicationContext context = new DefaultApplicationContext(Test1.class);
-        context.register(PropertyPathImporterTest.class);
-        PropertyPathImporterTest test = context.getBean(PropertyPathImporterTest.class);
-        Assert.assertEquals(12, test.age);
-    }
-
-    /**
-     * 使用文件路径读取
-     */
-    @Test
-    public void test2()
-    {
-        ApplicationContext context = new DefaultApplicationContext(Test2.class);
-        context.register(PropertyPathImporterTest.class);
-        PropertyPathImporterTest test = context.getBean(PropertyPathImporterTest.class);
-        Assert.assertEquals(12, test.age);
     }
 }
