@@ -2,8 +2,8 @@ package com.jfirer.jfire.test.function.aliastest;
 
 import com.jfirer.baseutil.bytecode.support.AnnotationContextFactory;
 import com.jfirer.baseutil.bytecode.support.SupportOverrideAttributeAnnotationContextFactory;
-import com.jfirer.jfire.core.DefaultApplicationContext;
 import com.jfirer.jfire.core.ApplicationContext;
+import com.jfirer.jfire.core.DefaultApplicationContext;
 import com.jfirer.jfire.core.prepare.annotation.ComponentScan;
 import com.jfirer.jfire.core.prepare.annotation.configuration.Configuration;
 import org.junit.Assert;
@@ -13,8 +13,7 @@ import javax.annotation.Resource;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.*;
 
 @Testalis3(t = "sada")
 public class AliasTest
@@ -40,19 +39,18 @@ public class AliasTest
         resource = annotationContextFactory.get(field).getAnnotation(Resource.class);
         assertEquals("demo", resource.name());
     }
+    @Test
+    public void test2()
+    {
+        ApplicationContext context = new DefaultApplicationContext(aliasCompopntScan.class);
+        SingleDemo         demo    = context.getBean("demo");
+        assertNotNull(demo);
+    }
 
     @ComponentScan("com.jfirer.jfire.test.function.aliastest")
     @Configuration
     public static class aliasCompopntScan
     {
 
-    }
-
-    @Test
-    public void test2()
-    {
-        ApplicationContext context = new DefaultApplicationContext(aliasCompopntScan.class);
-        SingleDemo         demo    = context.getBean("demo");
-        assertFalse(demo == null);
     }
 }
