@@ -9,21 +9,9 @@ import org.junit.Assert;
 public class Enhance
 {
 
-    @EnhanceClass(value = "com.jfirer.jfire.*.aop.Person", order = 2)
-    public static class EnhanceForOrder
-    {
-        public static String result = "";
-
-        @Before(value = "order()")
-        public void order2(ProceedPoint point)
-        {
-            result += "EnhanceForOrder_";
-        }
-    }
-
     private int    order;
     private String param;
-
+    private Object result;
     public int getOrder()
     {
         return order;
@@ -54,20 +42,15 @@ public class Enhance
     {
         EnhanceForOrder.result += "enhance";
     }
-
-    private Object result;
-
     public Object getResult()
     {
         return result;
     }
-
     @AfterReturning("order2(String,int)")
     public void order22(ProceedPoint point)
     {
         result = point.getResult();
     }
-
     @Around("myName(String)")
     public void testMyname(ProceedPoint point)
     {
@@ -75,7 +58,6 @@ public class Enhance
         point.invoke();
         System.out.println("环绕增强后");
     }
-
     @Around("testForVoidReturn()")
     public void testForVoidReturn(ProceedPoint point)
     {
@@ -84,23 +66,32 @@ public class Enhance
         point.invoke();
         System.out.println("环绕增强后");
     }
-
     @AfterThrowable("throwe(*)")
     public void throwe(ProceedPoint point)
     {
         System.out.println("dada");
         Verify.equal("aaaa", point.getE().getMessage(), "捕获到正确的异常");
     }
-
     @Before("getHomeName()")
     public void home1(ProceedPoint point)
     {
         System.out.println("before getHomeName");
     }
-
     @Before("getHome()")
     public void home2(ProceedPoint point)
     {
         System.out.println("before getHome");
+    }
+
+    @EnhanceClass(value = "com.jfirer.jfire.*.aop.Person", order = 2)
+    public static class EnhanceForOrder
+    {
+        public static String result = "";
+
+        @Before(value = "order()")
+        public void order2(ProceedPoint point)
+        {
+            result += "EnhanceForOrder_";
+        }
     }
 }
