@@ -17,6 +17,45 @@ import javax.annotation.Resource;
 @Import({HouseProvider.class, Data.NameProperty.class})
 public class Data
 {
+    @Resource(name = "house")
+    private House house;
+    @Resource(name = "house2")
+    private House house2;
+    @PropertyRead("name")
+    private String name;
+    @Bean
+    public Person person()
+    {
+        return new Person();
+    }
+    @Bean
+    @Conditional(Person2Condition.class)
+    public Person person2()
+    {
+        return new Person();
+    }
+    @Bean
+    public Person person4()
+    {
+        Person person = new Person();
+        person.setName(name);
+        return person;
+    }
+    @Bean
+    public Person person5()
+    {
+        Person person = new Person();
+        person.setName(house.name());
+        return person;
+    }
+    @Bean
+    public Person person7()
+    {
+        Person person = new Person();
+        person.setName(house2.name());
+        return person;
+    }
+
     public static class NameProperty implements ContextPrepare
     {
 
@@ -32,50 +71,5 @@ public class Data
         {
             return 0;
         }
-    }
-
-    @Resource(name = "house")
-    private House house;
-    @Resource(name = "house2")
-    private House house2;
-
-    @Bean
-    public Person person()
-    {
-        return new Person();
-    }
-
-    @PropertyRead("name")
-    private String name;
-
-    @Bean
-    @Conditional(Person2Condition.class)
-    public Person person2()
-    {
-        return new Person();
-    }
-
-    @Bean
-    public Person person4()
-    {
-        Person person = new Person();
-        person.setName(name);
-        return person;
-    }
-
-    @Bean
-    public Person person5()
-    {
-        Person person = new Person();
-        person.setName(house.name());
-        return person;
-    }
-
-    @Bean
-    public Person person7()
-    {
-        Person person = new Person();
-        person.setName(house2.name());
-        return person;
     }
 }
