@@ -1,6 +1,5 @@
 package com.jfirer.jfire.test.function.beanannotest;
 
-import com.jfirer.baseutil.bytecode.support.AnnotationContext;
 import com.jfirer.baseutil.bytecode.support.AnnotationContextFactory;
 import com.jfirer.jfire.core.ApplicationContext;
 import com.jfirer.jfire.core.DefaultApplicationContext;
@@ -20,18 +19,11 @@ public class MyBeanImport
     public Person importBean()
     {
         AnnotationContextFactory annotationContextFactory = DefaultApplicationContext.ANNOTATION_CONTEXT_FACTORY;
-        Optional<Person> any = context.getAllBeanRegisterInfos().stream()
-                                      .filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType())
-                                                                                          .isAnnotationPresent(Configuration.class))
-                                      .filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType())
-                                                                                          .isAnnotationPresent(MyImport.class))
-                                      .map(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType())
-                                                                                       .getAnnotation(MyImport.class))
-                                      .map(myImport -> {
-                                          Person person = new Person();
-                                          person.setName(myImport.name());
-                                          return person;
-                                      }).findAny();
+        Optional<Person> any = context.getAllBeanRegisterInfos().stream().filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(Configuration.class)).filter(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).isAnnotationPresent(MyImport.class)).map(beanRegisterInfo -> annotationContextFactory.get(beanRegisterInfo.getType()).getAnnotation(MyImport.class)).map(myImport -> {
+            Person person = new Person();
+            person.setName(myImport.name());
+            return person;
+        }).findAny();
         return any.orElse(null);
 //        for (Class<?> each : context.getConfigurationClassSet())
 //        {
