@@ -35,14 +35,12 @@ public class Person
     {
         this.name = name;
     }
-
-    public void setAge(int age)
-    {
-        this.age = age;
-    }
-
     public int getAge()
     {
         return age;
+    }
+    public void setAge(int age)
+    {
+        this.age = age;
     }
 }
